Various cleanups
- YAML formatting (yamllint)
- Pylint
- Pull Requests to separate Helm Repo (helm-test)
- Rustdoc will now error on warnings
