+++ .github/workflows/run.yml
@@ -1,3 +1,4 @@
+---
 name: Rust
 
 on:
