+++ template/.github/dependabot.yml
@@ -2,6 +2,7 @@
 # This file is automatically generated from the templates in stackabletech/operator-templating
 # DON'T MANUALLY EDIT THIS FILE
 # =============
+---
 version: 2
 updates:
   - package-ecosystem: "github-actions"
