+++ template/.github/workflows/daily_security.yml
@@ -2,11 +2,12 @@
 # This file is automatically generated from the templates in stackabletech/operator-templating
 # DON'T MANUALLY EDIT THIS FILE
 # =============
+---
 name: Security audit
 
 on:
   schedule:
-    - cron: '0 0 * * *'
+    - cron: '15 4 * * *'
   workflow_dispatch:
 
 jobs:
