+++ template/.github/workflows/publish_main_artifacts.yml.j2
@@ -2,7 +2,8 @@
 # This file is automatically generated from the templates in stackabletech/operator-templating
 # DON'T MANUALLY EDIT THIS FILE
 # =============
-name: Publish nightly artifacts from main
+---
+name: Publish nightly artifacts from main branch
 
 on:
   push:
@@ -11,15 +12,14 @@ on:
   schedule:
     - cron: '30 4 * * *'
   workflow_dispatch:
- 
 
 env:
   PRODUCT_NAME: {[ operator.product_string }]
   CARGO_TERM_COLOR: always
   CARGO_INCREMENTAL: '0'
   CARGO_PROFILE_DEV_DEBUG: '0'
-  RUSTFLAGS: "-D warnings -W rust-2021-compatibility"
-  REPO_HELM_DEV_URL: https://repo.stackable.tech/repository/helm-dev
+  RUSTFLAGS: "-D warnings"
+  REPO_HELM_URL: https://repo.stackable.tech/repository/helm-dev
   CHART_TESTING_CONFIG: deploy/helm/chart-testing.yaml
 
 jobs:
@@ -47,6 +47,7 @@ jobs:
 
       - name: Package Chart
         run: mkdir -p target/helm && helm package --destination target/helm deploy/helm/${{ env.PRODUCT_NAME }}-operator
+
       - name: Publish Chart
         env:
           NEXUS_PASSWORD: ${{ secrets.NEXUS_PASSWORD }}
@@ -55,6 +56,5 @@ jobs:
           /usr/bin/curl
           --fail
           -u 'github:${{ secrets.NEXUS_PASSWORD }}'
-          --upload-file "./$(find  target/helm/ -name *.tgz)"
-          "${{ env.REPO_HELM_DEV_URL }}/"
-
+          --upload-file "./$(find  target/helm/ -name '*.tgz')"
+          "${{ env.REPO_HELM_URL }}/"
