+++ template/.github/workflows/publish_pr_artifacts.yml.j2
@@ -2,7 +2,8 @@
 # This file is automatically generated from the templates in stackabletech/operator-templating
 # DON'T MANUALLY EDIT THIS FILE
 # =============
-name: Publish pull-requests artifacts to the dev repository
+---
+name: Publish pull-request artifacts
 
 on:
   pull_request:
@@ -12,8 +13,8 @@ env:
   CARGO_TERM_COLOR: always
   CARGO_INCREMENTAL: '0'
   CARGO_PROFILE_DEV_DEBUG: '0'
-  RUSTFLAGS: "-D warnings -W rust-2021-compatibility"
-  REPO_HELM_DEV_URL: https://repo.stackable.tech/repository/helm-dev
+  RUSTFLAGS: "-D warnings"
+  REPO_HELM_URL: https://repo.stackable.tech/repository/helm-test
   CHART_TESTING_CONFIG: deploy/helm/chart-testing.yaml
 
 jobs:
@@ -36,7 +37,7 @@ jobs:
           python-version: '3.x'
 
       - run: pip install -r ./python/requirements.txt
-      - run: python ./python/cargo-version.py -m mr${{ github.event.number }}
+      - run: python ./python/cargo_version.py -m pr${{ github.event.number }}
 
       - name: Build Docker image
         env:
@@ -49,6 +50,7 @@ jobs:
 
       - name: Package Chart
         run: mkdir -p target/helm && helm package --destination target/helm deploy/helm/${{ env.PRODUCT_NAME }}-operator
+
       - name: Publish Chart
         env:
           NEXUS_PASSWORD: ${{ secrets.NEXUS_PASSWORD }}
@@ -57,6 +59,5 @@ jobs:
           /usr/bin/curl
           --fail
           -u 'github:${{ secrets.NEXUS_PASSWORD }}'
-          --upload-file "./$(find  target/helm/ -name *.tgz)"
-          "${{ env.REPO_HELM_DEV_URL }}/"
-
+          --upload-file "./$(find  target/helm/ -name '*.tgz')"
+          "${{ env.REPO_HELM_URL }}/"
