+++ template/.github/workflows/publish_release_artifacts.yml.j2
@@ -2,7 +2,8 @@
 # This file is automatically generated from the templates in stackabletech/operator-templating
 # DON'T MANUALLY EDIT THIS FILE
 # =============
-name: Publish-Release-Artifacts
+---
+name: Publish release artifacts
 
 on:
   push:
@@ -14,9 +15,9 @@ env:
   CARGO_TERM_COLOR: always
   CARGO_INCREMENTAL: '0'
   CARGO_PROFILE_DEV_DEBUG: '0'
-  RUSTFLAGS: "-D warnings -W rust-2021-compatibility"
-  REPO_HELM_STABLE_URL: https://repo.stackable.tech/repository/helm-stable
-  CT_CONFIG: deploy/helm/ct.yaml
+  RUSTFLAGS: "-D warnings"
+  REPO_HELM_URL: https://repo.stackable.tech/repository/helm-stable
+  CHART_TESTING_CONFIG: deploy/helm/chart-testing.yaml
 
 jobs:
   helm:
@@ -52,5 +53,5 @@ jobs:
           /usr/bin/curl
           --fail
           -u 'github:${{ secrets.NEXUS_PASSWORD }}'
-          --upload-file "./$(find  target/helm/ -name *.tgz)"
-          "${{ env.REPO_HELM_STABLE_URL }}/"
+          --upload-file "./$(find  target/helm/ -name '*.tgz')"
+          "${{ env.REPO_HELM_URL }}/"
