+++ template/.github/workflows/rust.yml
@@ -2,7 +2,8 @@
 # This file is automatically generated from the templates in stackabletech/operator-templating
 # DON'T MANUALLY EDIT THIS FILE
 # =============
-name: Rust
+---
+name: Rust checks
 
 on:
   push:
@@ -17,6 +18,8 @@ env:
   CARGO_INCREMENTAL: '0'
   CARGO_PROFILE_DEV_DEBUG: '0'
   RUSTFLAGS: "-D warnings"
+  RUSTDOCFLAGS: "-D warnings"
+  RUST_LOG: "info"
 
 jobs:
 
@@ -79,10 +82,10 @@ jobs:
       - uses: actions/checkout@v2.4.0
       - uses: actions-rs/toolchain@v1.0.7
         with:
-            profile: minimal
-            toolchain: stable
-            components: clippy
-            override: true
+          profile: minimal
+          toolchain: stable
+          components: clippy
+          override: true
       - uses: Swatinem/rust-cache@v1.3.0
         with:
           key: clippy
