+++ template/deploy/DO_NOT_EDIT.md
@@ -1,0 +1,4 @@
+These Helm charts and manifests are automatically generated.
+Please do not edit anything in this directory manually.
+
+TODO: Describe how they are generated
