+++ template/deploy/helm/[[product]]-operator/Chart.yaml.j2
@@ -1,33 +1,11 @@
-# =============
-# This file is automatically generated from the templates in stackabletech/operator-templating
-# DON'T MANUALLY EDIT THIS FILE
-# =============
-
+---
 apiVersion: v2
 name: {[ operator.product_string }]-operator
-description: The Stackable Operator for {[ operator.pretty_string }]
+version: <version_placeholder>
+appVersion: "<version_placeholder>"
 
-# A chart can be either an 'application' or a 'library' chart.
-#
-# Application charts are a collection of templates that can be packaged into versioned archives
-# to be deployed.
-#
-# Library charts provide useful utilities or functions for the chart developer. They're included as
-# a dependency of application charts to inject those utilities and functions into the rendering
-# pipeline. Library charts do not define any templates and therefore cannot be deployed.
-type: application
+home: https://github.com/stackabletech/{[ operator.product_string }]-operator
 
 maintainers:
-  - name: stackable
-    url: github.com/stackabletech
-
-# This is the chart version. This version number should be incremented each time you make changes
-# to the chart and its templates, including the app version.
-# Versions are expected to follow Semantic Versioning (https://semver.org/)
-version: <version_placeholder>
-
-# This is the version number of the application being deployed. This version number should be
-# incremented each time you make changes to the application. Versions are not expected to
-# follow Semantic Versioning. They should reflect the version the application is using.
-# It is recommended to use it with quotes.
-appVersion: <version_placeholder>
+  - name: Stackable
+    url: https://www.stackable.tech
