+++ template/deploy/helm/[[product]]-operator/README.md.j2
@@ -1,8 +1,3 @@
-[//]: # (=============)
-[//]: # (This file is automatically generated from the templates in stackabletech/operator-templating)
-[//]: # (DON'T MANUALLY EDIT THIS FILE)
-[//]: # (=============)
-
 # Helm Chart for Stackable Operator for {[ operator.pretty_string }]
 
 This Helm Chart can be used to install Custom Resource Definitions and the Operator for {[ operator.pretty_string }] provided by Stackable.
@@ -14,7 +9,7 @@ This Helm Chart can be used to install Custom Resource Definitions and the Opera
 - Install [Helm](https://helm.sh/docs/intro/install/)
 
 
-## Install the Stackble Operator for {[ operator.pretty_string }]
+## Install the Stackable Operator for {[ operator.pretty_string }]
 
 ```bash
 # From the root of the operator repository
@@ -24,15 +19,11 @@ helm install {[ operator.product_string }]-operator deploy/helm/{[ operator.prod
 ```
 
 
+## Usage of the CRDs
 
+The usage of this operator and its CRDs is described in the [documentation](https://docs.stackable.tech/{[ operator.product_string }]/index.html)
 
-## Create a {[ operator.pretty_string }] Cluster
-
-as described [here](https://docs.stackable.tech/{[ operator.product_string }]/index.html)
-
-
-
-The operator has example requests included in the [`/examples`](https://github.com/stackabletech/{[ operator.product_string }]/operator/tree/main/examples) directory that can be used to spin up a cluster.
+The operator has example requests included in the [`/examples`](https://github.com/stackabletech/{[ operator.product_string }]/operator/tree/main/examples) directory.
 
 
 ## Links
