+++ template/deploy/helm/[[product]]-operator/templates/configmap.yaml
@@ -1,7 +1,4 @@
-# =============
-# This file is automatically generated from the templates in stackabletech/operator-templating
-# DON'T MANUALLY EDIT THIS FILE
-# =============
+---
 apiVersion: v1
 data:
 {{ (.Files.Glob "configs/*").AsConfig | indent 2 }}
