+++ template/deploy/helm/[[product]]-operator/templates/deployment.yaml.j2
@@ -1,7 +1,4 @@
-# =============
-# This file is automatically generated from the templates in stackabletech/operator-templating
-# DON'T MANUALLY EDIT THIS FILE
-# =============
+---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
