+++ template/deploy/helm/[[product]]-operator/templates/roles.yaml.j2
@@ -1,3 +1,4 @@
+---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
