+++ template/deploy/helm/[[product]]-operator/templates/serviceaccount.yaml.j2
@@ -1,7 +1,4 @@
-# =============
-# This file is automatically generated from the templates in stackabletech/operator-templating
-# DON'T MANUALLY EDIT THIS FILE
-# =============
+---
 {{ if .Values.serviceAccount.create -}}
 apiVersion: v1
 kind: ServiceAccount
