+++ template/deploy/helm/[[product]]-operator/values.yaml.j2
@@ -1,11 +1,5 @@
-# =============
-# This file is automatically generated from the templates in stackabletech/operator-templating
-# DON'T MANUALLY EDIT THIS FILE
-# =============
-
 # Default values for {[ operator.product_string }]-operator.
-# This is a YAML-formatted file.
-# Declare variables to be passed into your templates.
+---
 image:
   repository: docker.stackable.tech/stackable/{[ operator.product_string }]-operator
   pullPolicy: IfNotPresent
