+++ template/deploy/helm/chart_testing.yaml
@@ -1,3 +1,4 @@
+---
 remote: origin
 target-branch: main
 chart-dirs:
